# shared/openapi/server-api.yaml
openapi: 3.1.0
info:
  title: Power Monitor Server API
  description: Central server for pm-protocol/1.0.0 sensor fleets
  version: 1.0.0
paths:
...
  /api/v1/firmware-releases:
    get:
      tags:
      - firmware
      summary: List Releases
      operationId: list_releases_api_v1_firmware_releases_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Releases Api V1 Firmware Releases Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - firmware
      summary: Upload Release
      operationId: upload_release_api_v1_firmware_releases_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_release_api_v1_firmware_releases_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Release Api V1 Firmware Releases Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/firmware-deployments:
    post:
      tags:
      - firmware
      summary: Create Deployments
      operationId: create_deployments_api_v1_firmware_deployments_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirmwareDeploymentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Deployments Api V1 Firmware Deployments Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - firmware
      summary: List Deployments
      operationId: list_deployments_api_v1_firmware_deployments_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Deployments Api V1 Firmware Deployments Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/device-firmware/{release_id}/download:
    get:
      tags:
      - firmware
      summary: Device Download Firmware
      operationId: device_download_firmware_api_v1_device_firmware__release_id__download_get
      parameters:
      - name: release_id
        in: path
        required: true
        schema:
          type: string
          title: Release Id
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_release_api_v1_firmware_releases_post:
      properties:
        manifest_json:
          type: string
          title: Manifest Json
        public_key_pem:
          type: string
          title: Public Key Pem
        binary:
          type: string
          contentMediaType: application/octet-stream
          title: Binary
      type: object
      required:
      - manifest_json
      - public_key_pem
      - binary
      title: Body_upload_release_api_v1_firmware_releases_post
...
    FirmwareDeploymentCreate:
      properties:
        firmware_release_id:
          type: string
          title: Firmware Release Id
        device_ids:
          items:
            type: string
          type: array
          maxItems: 1000
          minItems: 1
          title: Device Ids
        scheduled_at:
          type: string
          format: date-time
          title: Scheduled At
      additionalProperties: false
      type: object
      required:
      - firmware_release_id
      - device_ids
      - scheduled_at
      title: FirmwareDeploymentCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError